Move SpaceSelectMenu into separete component
